feat: add node list to entities/producers card
