+++ webapp/src/language/en.lacchain.json
@@ -45,20 +45,6 @@
     "transactions": "Transactions"
   },
   "dashboardNodes": {
-    "name": "Name",
-    "endpoints": "Endpoints",
-    "keys": "Keys",
-    "features": "Features",
-    "producer": "Entity",
-    "nodeType": "Node type",
-    "type": "Type",
-    "description": "Description",
-    "color": "Color",
-    "account": "Account",
-    "website": "Website",
-    "votes": "Votes",
-    "rewards": "Rewards",
-    "producerCountry": "Entity country",
     "htmlTitle": "Node Distribution - Dashboard"
   },
   "dashboardRewards": {
@@ -119,7 +105,8 @@
     "social": "Social",
     "ownershipDisclosure": "Ownership Disclosure",
     "chainResources": "Chain Resources",
-    "otherResources": "Other Resources"
+    "otherResources": "Other Resources",
+    "nodes": "Nodes"
   },
   "nodeSummary": {
     "nodeName": "Node name",
