+++ webapp/src/language/es.lacchain.json
@@ -43,20 +43,6 @@
     "transactions": "Transacciones"
   },
   "dashboardNodes": {
-    "name": "Nombre",
-    "endpoints": "Endpoints",
-    "keys": "Keys",
-    "features": "Caracteristicas",
-    "producer": "Entidad",
-    "nodeType": "Tipo de nodo",
-    "type": "Tipo",
-    "description": "Descripción",
-    "color": "Color",
-    "account": "Cuenta",
-    "website": "Sitio web",
-    "votes": "Votos",
-    "rewards": "Recompensas",
-    "producerCountry": "País de la entidad",
     "htmlTitle": "Distribución de nodos - Dashboard"
   },
   "dashboardRewards": {
@@ -117,7 +103,8 @@
     "social": "Social",
     "ownershipDisclosure": "Divulgación de propiedad",
     "chainResources": "Recursos de la cadena",
-    "otherResources": "Otros recursos"
+    "otherResources": "Otros recursos",
+    "nodes": "Nodos"
   },
   "nodeSummary": {
     "nodeName": "Nombre del nodo",
